Async client design overhaul.
First, TFuture was split into two parts, a TFuture interface for accessi
ng the result, and a TPromise implementation for actually setting it. Se
cond, the relation between TAsyncClient/TAsyncSocketClient was overhaule
d so that it is actually extensible if other I/O devices are ever added.

The commit seems fairly big, and indeed it could have been split into
two parts, but most of the diff are whitespace changes (due to indentati
on).