Don't use contracts for checking isOpen() for transports.
This effectively made it impossible to close() transports from other methods as a result of failed operations, because it would mean that TTransport.isOpen would have to be checked at every call site to avoid possibly violating the contracts. TSocket was already broken in this regard, and the TAsyncSocket implementation would have been unnecessarily complicated as well. Additionally, although this will eventually be resolved, code generation for inherited contracts is currently broken with DMD, forcing all the checks to be disabled to avoid crashes.

The new/clarified semantics for isOpen are admittedly not to pretty, but there is really no good alternative.