Make requirement default depending on field type.
Previously, OPT_IN_REQ_OUT was used as default in the codegen templates
if no requirement was explicitly specified. This makes sense, because it
is also the default of the Thrift compiler, but made serializing
pre-existing structs harder, because the code looked for isSet fields
for all the non-nullable members.  Now, the default is REQUIRED/OPTIONAL
depending on whether a field is mutable, and OPT_IN_REQ_OUT is
explicitly specified by the generator.