Refactored SSL context handling.
Now, a TSSLContext is used that is passed to new sockets instead of TSSLSocketFactory, which used to create new sockets. This is so that other parts of the library can use SSL support as well without being referenced from thrift.transport.ssl.

Also includes some minor cleanups made in the process.